refactor: consolidate dir-glob exclude-merge factory in owned-paths
The markdownlint `ignores` merge hand-rolled merge/unmerge logic that was
near-identical to dirGlobExcludeMerge (dprint/oxfmt) — only the field name
and glob form differed. Generalize dirGlobExcludeMerge with an optional
per-dir glob builder (defaulting to the existing `<dir>/**` form so
dprint/oxfmt are byte-identical) and move it from the typescript pack to
owned-paths.ts, beside SAFEWORD_IGNORE_DIRS, the data it consumes. This
removes the duplicated boilerplate and the cross-pack coupling of a
language-agnostic helper. Behavior-preserving; dprint/oxfmt/markdownlint
reconcile tests unchanged and green.
